feat: implement Side Panel route tutorial
