+++ src/components/layout/sidePannel/EventCardView.module.css
