+++ src/components/tutorial/guides/tutorialRegistry.ts
@@ -1,0 +1,10 @@
+import { DETAIL_VIEW_GUIDE } from "./detailViewGuide";
+import { MAIN_ROUTE_GUIDE } from "./mainRouteGuide";
+import { SIDE_PANEL_RESIZE_GUIDE } from "./sidePanelResizeGuide";
+import type { GuideDefinitions } from "../types";
+
+export const TUTORIAL_GUIDES: GuideDefinitions[] = [
+	SIDE_PANEL_RESIZE_GUIDE,
+	DETAIL_VIEW_GUIDE,
+	MAIN_ROUTE_GUIDE,
+];
